+++ ios/Classes/SwiftAudioWaveformsPlugin.swift
@@ -27,8 +27,11 @@ public class SwiftAudioWaveformsPlugin: NSObject, FlutterPlugin {
         let args = call.arguments as? Dictionary<String, Any>
         switch call.method {
         case Constants.startRecording:
-            audioRecorder.startRecording(result,  args?[Constants.path] as? String,
-                                         args?[Constants.encoder] as? Int, args?[Constants.sampleRate] as? Int, args?[Constants.bitRate] as? Int,Constants.fileNameFormat, args?[Constants.useLegacyNormalization] as? Bool,overrideAudioSession: (args?[Constants.overrideAudioSession] as? Bool) ?? true )
+            guard let args = call.arguments as? Dictionary<String, Any> else {
+                result(FlutterError(code: Constants.audioWaveforms, message: "Invalid Arguments", details: nil))
+                return
+            }
+            audioRecorder.startRecording(result, RecordingSettings.fromJson((args)))
             break
         case Constants.pauseRecording:
             audioRecorder.pauseRecording(result)